pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ppa-agent"
version = "0.1.0"
description = "AI Agent for PPA insurance new business quoting process"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12.0"
typing-extensions = "^4.13.2"
openai = "^1.74.0"
tiktoken = "^0.9.0"
tenacity = "^9.1.2"
SQLAlchemy = "^2.0.40"
PyYAML = "^6.0.2"
requests = "^2.32.3"
orjson = "^3.10.16"
ormsgpack = "^1.9.1"
xxhash = "^3.5.0"
jsonpatch = "^1.33"
packaging = "^24.2"
tqdm = "^4.67.1"
regex = "^2024.11.6"
langgraph = {extras = ["dev"], version = "^0.3.31"}
python-dotenv = "^1.1.0"
pydantic = "^2.11.3"
httpx = "^0.28.1"
google-genai = "^1.10.0"
zstandard = "^0.23.0"
d3graph = "^2.5.0"
langgraph-sdk = "^0.1.61"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.1"}
langserve = "^0.3.1"
sse-starlette = "~2.1.0"
langchain = "^0.3.23"
langchain-core = "^0.3.54"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
pre-commit = "^4.2.0"
langgraph-cli = {extras = ["inmem"], version = "^0.2.5"}

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "8.1"
addopts = "-ra -q"
testpaths = [
    "tests",
]
